refactor: split exit processor core tests
refactor: split piggyback-related tests

refactor: split exit tests: generic, standard exit, finalization, canon.



refactor: final tidies + liberate plain SE validity tests from OMG.State
